geoguess_project/backend/requirements.txt
flask
flask-cors
torch
numpy
pandas
scikit-learn
Pillow
tqdm
clip-by-openai